pyproject.toml
[tool.poetry]
name = "nonebot-plugin-templates"
version = "0.1.3"
description = "为nonebot提供一系列jinja2 template模板,方便的生成图片"
authors = ["canxin <1969730106@qq.com>"]
readme = "README.md"
packages = [{ include = "nonebot_plugin_templates" }]

[tool.poetry.dependencies]
python = "^3.8"
nonebot-plugin-htmlrender = "^0.2.1"
nonebot2 = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
